% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-integers.R
\name{insert_assign_random_set_seed}
\alias{insert_assign_random_set_seed}
\title{Insert \code{set.seed()} and bind to name \code{SEED}}
\usage{
insert_assign_random_set_seed(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=get_random_integer]{get_random_integer()}}}
}
\value{
A \code{set.seed()} string
}
\description{
Insert \code{set.seed()} and bind to name \code{SEED}
}
\examples{
\dontrun{
insert_assign_random_set_seed()
}
}
