% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-integers.R
\name{insert_random_integer}
\alias{insert_random_integer}
\title{Insert random integer at cursor}
\usage{
insert_random_integer(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=get_random_integer]{get_random_integer()}}}
}
\value{
A string representing a randomly sampled integer
}
\description{
Insert random integer at cursor
}
